implemented user cache register access
nominally ten registers supported, 0--9. getop returns a flag of
'VARIABLE' (maybe better would be SYM_CONST or a special REGISTER) when
it encounters a $ in the input and stores $d in the string. this is
passed to a function in stack, getcache, that returns the cache given
this string. untested.