implement whitespace consume macro
consuming white space from a stream is frequently done and so should be
put in a macro. white space may be delimiting or meaningless, for example,
the below postfix expression has two meaningless whitespace parts and
three delimiting

A = 3 5 A *
 ^ ^

so in the variable assignment parsing, one likely would use the
whitespace consume macro. in the case of redundant whitespace in
delimitation, it also has to be consumed, like

A = 3 \t5 \t\tA \t*