stdin is put into input buffer for main calc
this doubles memory requirements for the system, since stdin is
presumably not deallocated for some other purpose when you consume with
getchar. but for testing purposes a buffer input to the tokenizer is
valuable, and it is easily changed to buffer in increments later for
efficiency (in getch, load only 10 chars or so, then when one encounters
EOF, check if standard in is EOF and if not load up to 10 chars from
standard in).

buffer is initialized/flushed to EOF, not NULL, to stop the input.